rename : Controller 코드 REST API, View directory 구분 및 코드 정렬
