+++ src/main/java/com/example/classicjeans/controller/view/NursingHomeViewController.java
@@ -1,4 +1,4 @@
-package com.example.classicjeans.controller;
+package com.example.classicjeans.controller.view;
 
 import com.example.classicjeans.dto.response.NursingHomeResponse;
 import com.example.classicjeans.service.NursingHomeService;
