+++ src/main/java/com/example/classicjeans/controller/view/UserViewController.java
@@ -1,4 +1,4 @@
-package com.example.classicjeans.controller;
+package com.example.classicjeans.controller.view;
 
 import com.example.classicjeans.dto.request.UsersRequest;
 import com.example.classicjeans.dto.response.FamilyInfoResponse;
